refactor initial commit
