+++ classes/main.rb
@@ -1,53 +1,8 @@
 require_relative 'app'
-app = App.new
+require_relative 'operation'
 
-def display(app)
-  puts "
-            [1] -> List all books
-            [2] -> List all people
-            [3] -> Create a person
-            [4] -> Create a book
-            [5] -> Creat a rental
-            [6] -> List all rentals for a given person id
-            [7] -> quit
-            Enter Selection:
-        "
-  option = gets.chomp.to_i
-
-  key_select(option, app)
-end
-
-# rubocop:disable Metrics/CyclomaticComplexity
-def key_select(option, app)
-  case option
-  when 1
-    app.list_books
-
-  when 2
-    app.list_people
-
-  when 3
-    app.create_person
-
-  when 4
-    app.create_book
-
-  when 5
-    app.create_rental
-
-  when 6
-    app.list_rental
-
-  when 7
-    app.quit_app
-  else
-    puts 'enter valid input'
-  end
-  display(app) if option < 7
-end
-# rubocop:enable Metrics/CyclomaticComplexity
-
-def main(app)
+def main
+  app = App.new
   display(app)
 end
-main(app)
+main
